% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgamglm.R
\name{bgamglm}
\alias{bgamglm}
\title{bgamglm}
\usage{
bgamglm(w, Z, pm, pv, p0, tol)
}
\arguments{
\item{w}{vector of responses}

\item{Z}{design matrix}

\item{pm}{prior mean}

\item{pv}{prior covariance matrix}

\item{p0}{starting guess for mode}

\item{tol}{tolerance for stopping}
}
\description{
bgamglm code from: "Variational approximation for heteroscedastic models and matching
pursuit algorithms" by David Nott et al.
}
\author{
Minh-Ngoc Tran (ngoctm@nus.edu.sg) with a helper function from David Nott
}
